+++ example/index.php
@@ -8,19 +8,19 @@
 
 require_once __DIR__ . '/../vendor/autoload.php';
 
-$response = esolat()
-    ->timeline()
-    ->zone('PNG01')
-    ->displayAs(1) // must be 1
-    ->setDate('2018-10-10') // if this is not set, it will automatically get current date
-    ->fetch();
-//
 //$response = esolat()
 //    ->timeline()
 //    ->zone('PNG01')
-//    ->displayAs(2) // default is 2 (Week)
+//    ->displayAs(1) // must be 1
+//    ->setDate('2018-10-10') // if this is not set, it will automatically get current date
 //    ->fetch();
 //
+$response = esolat()
+    ->timeline()
+    ->zone('PNG01')
+    ->displayAs(2) // default is 2 (Week)
+    ->fetch();
+
 //$data = esolat()
 //    ->timeline()
 //    ->zone('PNG01')
@@ -37,6 +37,6 @@ $response = esolat()
 //    ->fetch();
 
 //$response = esolat()->location_list('negeri sembilan');
-
+//$response = esolat()->date_to_hijri(\Carbon\Carbon::now());
 header('Content-type: application/json');
 echo json_encode($response, JSON_PRETTY_PRINT);