+++ changelog.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## 1.1 Add Hijri Dates and Converter
+- Add Hijri Date Converter
+- Add Hijri Date to Waktu Solat JSON
+- Extend Carbon as IslamicCarbon
+- Fix Readme and Sample
+- Fix Tests
+
 ## 1.0 First Release
 - Add E-solat Jakim Timeline
 - Features: Day, Week, Month, Year, Fetch Location in Malaysia